Small QoL for the backend. Should make the mod a bit more performant.
