+++ gradle/libs.versions.toml
@@ -1,7 +1,7 @@
 [versions]
 # plugins
-fabric_loom = "1.0-SNAPSHOT"
-quilt_flower = "1.7.4"
+fabric_loom = "1.+"
+quilt_flower = "1.+"
 
 # game dependencies https://modmuss50.me/fabric.html and for quilt mappings check: https://github.com/QuiltMC/quilt-mappings/tags
 minecraft="1.18.2"
@@ -10,11 +10,9 @@ fabric_api="0.67.0+1.18.2"
 
 # third party dependencies
 modmenu="3.2.3" # check https://www.curseforge.com/minecraft/mc-mods/modmenu/files/all?filter-game-version=1738749986%3a73250
-jxlib="0.1.6.5+1.18.2"
-cloth_config = "6.4.90"
 emi = "0.5.3+1.18.2"
 emi_loot="4159529" # check https://www.curseforge.com/minecraft/mc-mods/emi-loot/files/
-radiancelib = "0.1.3.1-alpha+1.18.2"
+radiancelib = "0.1.3.2-alpha+1.18.2"
 
 [libraries]
 minecraft = { module = "com.minecraft:minecraft", version.ref = "minecraft" }
@@ -25,8 +23,6 @@ fabric_api = { module = "net.fabricmc.fabric-api:fabric-api", version.ref = "fab
 
 # third party dependencies
 modmenu = { module = "com.terraformersmc:modmenu", version.ref = "modmenu" }
-jxlib = { module = "com.joostmsoftware:JXLib", version.ref = "jxlib" }
-cloth_config = { module = "me.shedaniel.cloth:cloth-config-fabric", version.ref = "cloth_config"}
 emi = { module = "dev.emi:emi", version.ref = "emi" }
 emi_loot = { module = "curse.maven:emi-loot-681783", version.ref = "emi_loot" }
 radiancelib = { module = "com.radiancemc:radiancelib", version.ref = "radiancelib" }
