+++ pfti-fabric/src/main/java/com/radiancemc/pfti/common/item/PftiItems.java
@@ -3,8 +3,7 @@ package com.radiancemc.pfti.common.item;
 import com.radiancemc.pfti.Pfti;
 import com.radiancemc.pfti.api.core.pet.PetItem;
 import com.radiancemc.pfti.api.core.pet.PetTypes;
-import com.radiancemc.pfti.common.item.pets.PetItems.*;
-import com.radiancemc.pfti.common.item.pets.StatusEffectPetItem;
+import com.radiancemc.pfti.common.item.pets.*;
 import net.minecraft.item.Item;
 import java.util.HashMap;
 import java.util.Map;
