+++ pfti-fabric/src/main/java/com/radiancemc/pfti/common/item/pets/SugarCanePetItem.java
@@ -1,28 +1,25 @@
-package com.radiancemc.pfti.common.item.pets.PetItems;
+package com.radiancemc.pfti.common.item.pets;
 
-import com.radiancemc.pfti.api.core.pet.PetItem;
 import com.radiancemc.pfti.api.core.pet.PetTypes;
 import net.minecraft.client.item.TooltipContext;
-import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.text.Text;
 import net.minecraft.text.TranslatableText;
-import net.minecraft.util.Formatting;
 import net.minecraft.world.World;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.List;
 
-public class TaterPetItem extends PetItem {
+public class SugarCanePetItem extends StatusEffectPetItem {
 
 
-    public TaterPetItem(int petTier, PetTypes type, Item.Settings settings) {
+    public SugarCanePetItem(int petTier, PetTypes type, Settings settings) {
         super(petTier, type, settings);
     }
 
     @Override
     public void appendTooltip(ItemStack stack, @Nullable World world, List<Text> tooltip, TooltipContext context) {
-        tooltip.add(new TranslatableText("taterpetitem.info").formatted(Formatting.GRAY));
+        tooltip.add(new TranslatableText("tooltip.sugarcanepetitem.info"));
         super.appendTooltip(stack, world, tooltip, context);
     }
 }
