+++ pfti-fabric/src/main/java/com/radiancemc/pfti/common/item/pets/TaterPetItem.java
@@ -1,26 +1,28 @@
-package com.radiancemc.pfti.common.item.pets.PetItems;
+package com.radiancemc.pfti.common.item.pets;
 
-import com.radiancemc.pfti.common.item.pets.StatusEffectPetItem;
+import com.radiancemc.pfti.api.core.pet.PetItem;
 import com.radiancemc.pfti.api.core.pet.PetTypes;
 import net.minecraft.client.item.TooltipContext;
+import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.text.Text;
 import net.minecraft.text.TranslatableText;
+import net.minecraft.util.Formatting;
 import net.minecraft.world.World;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.List;
 
-public class SugarCanePetItem extends StatusEffectPetItem {
+public class TaterPetItem extends PetItem {
 
 
-    public SugarCanePetItem(int petTier, PetTypes type, Settings settings) {
+    public TaterPetItem(int petTier, PetTypes type, Item.Settings settings) {
         super(petTier, type, settings);
     }
 
     @Override
     public void appendTooltip(ItemStack stack, @Nullable World world, List<Text> tooltip, TooltipContext context) {
-        tooltip.add(new TranslatableText("tooltip.sugarcanepetitem.info"));
+        tooltip.add(new TranslatableText("taterpetitem.info").formatted(Formatting.GRAY));
         super.appendTooltip(stack, world, tooltip, context);
     }
 }
